global sparsity pattern added

+++ tests/utils/helpers.hpp
@@ -6,11 +6,11 @@
 
 template <typename T>
     requires std::is_arithmetic_v<T>
-std::vector<T> generateRandomVector(int size)
+std::vector<T> generateRandomVector(size_t size)
 {
     std::vector<T> vector(size, T());
 
-    for (int i = 0; i < size; ++i)
+    for (size_t i = 0; i < size; ++i)
     {
         vector[i] = rand() % 101;
     }
@@ -20,13 +20,13 @@ std::vector<T> generateRandomVector(int size)
 
 template <typename T>
     requires std::is_arithmetic_v<T>
-std::vector<std::vector<T>> generateRandomMatrix(int rows, int cols)
+std::vector<std::vector<T>> generateRandomMatrix(size_t rows, size_t cols)
 {
     std::vector<std::vector<T>> matrix(rows, std::vector<T>(cols, T()));
 
-    for (int i = 0; i < rows; ++i)
+    for (size_t i = 0; i < rows; ++i)
     {
-        for (int j = 0; j < cols; ++j)
+        for (size_t j = 0; j < cols; ++j)
         {
             matrix[i][j] = rand() % 101;
         }
@@ -39,13 +39,13 @@ template <typename T>
     requires std::is_arithmetic_v<T>
 std::vector<std::vector<T>> addMatrices(const std::vector<std::vector<T>> &a, const std::vector<std::vector<T>> &b)
 {
-    int rows = a.size();
-    int cols = a.front().size();
+    size_t rows = a.size();
+    size_t cols = a.front().size();
     std::vector<std::vector<T>> result(rows, std::vector<T>(cols, T()));
 
-    for (int i = 0; i < rows; ++i)
+    for (size_t i = 0; i < rows; ++i)
     {
-        for (int j = 0; j < cols; ++j)
+        for (size_t j = 0; j < cols; ++j)
         {
             result[i][j] = a[i][j] + b[i][j];
         }
@@ -58,13 +58,13 @@ template <typename T>
     requires std::is_arithmetic_v<T>
 std::vector<std::vector<T>> subtractMatrices(const std::vector<std::vector<T>> &a, const std::vector<std::vector<T>> &b)
 {
-    int rows = a.size();
-    int cols = a.front().size();
+    size_t rows = a.size();
+    size_t cols = a.front().size();
     std::vector<std::vector<T>> result(rows, std::vector<T>(cols, T()));
 
-    for (int i = 0; i < rows; ++i)
+    for (size_t i = 0; i < rows; ++i)
     {
-        for (int j = 0; j < cols; ++j)
+        for (size_t j = 0; j < cols; ++j)
         {
             result[i][j] = a[i][j] - b[i][j];
         }
@@ -77,14 +77,14 @@ template <typename T>
     requires std::is_arithmetic_v<T>
 std::vector<T> multiplyMatrixByVector(const std::vector<std::vector<T>> &matrix, const std::vector<T> &vec)
 {
-    int rows = static_cast<int>(matrix.size());
-    int cols = static_cast<int>(matrix[0].size());
+    size_t rows = matrix.size();
+    size_t cols = matrix[0].size();
 
     std::vector<T> result(rows, T());
 
-    for (int i = 0; i < rows; ++i)
+    for (size_t i = 0; i < rows; ++i)
     {
-        for (int j = 0; j < cols; ++j)
+        for (size_t j = 0; j < cols; ++j)
         {
             result[i] += matrix[i][j] * vec[j];
         }
@@ -96,18 +96,18 @@ template <typename T>
     requires std::is_arithmetic_v<T>
 std::vector<std::vector<T>> multiplyMatrices(const std::vector<std::vector<T>> &a, const std::vector<std::vector<T>> &b)
 {
-    int rowsA = a.size();
-    int colsA = a.front().size();
-    int colsB = b.front().size();
+    size_t rowsA = a.size();
+    size_t colsA = a.front().size();
+    size_t colsB = b.front().size();
 
     std::vector<std::vector<T>> result(rowsA, std::vector<T>(colsB, T()));
 
-    for (int i = 0; i < rowsA; ++i)
+    for (size_t i = 0; i < rowsA; ++i)
     {
-        for (int j = 0; j < colsB; ++j)
+        for (size_t j = 0; j < colsB; ++j)
         {
             result[i][j] = T();
-            for (int k = 0; k < colsA; ++k)
+            for (size_t k = 0; k < colsA; ++k)
             {
                 result[i][j] += a[i][k] * b[k][j];
             }
@@ -121,7 +121,7 @@ template <typename T>
 std::ostream &operator<<(std::ostream &os, const std::vector<T> &vec)
 {
     os << "[";
-    for (int i = 0; i < static_cast<int>(vec.size()); ++i)
+    for (size_t i = 0; i < vec.size(); ++i)
     {
         if (i > 0)
         {
